tests: divide `polygon` into `point` and `box`
