+++ tests/bool2d/test_contains.py
@@ -21,9 +21,6 @@ from shapepy.geometry.jordancurve import JordanCurve
 @pytest.mark.order(23)
 @pytest.mark.dependency(
     depends=[
-        "tests/geometry/test_polygon.py::test_end",
-        "tests/geometry/test_jordan_polygon.py::test_end",
-        "tests/geometry/test_jordan_curve.py::test_end",
         "tests/bool2d/test_primitive.py::test_end",
     ],
     scope="session",
